Persist due date for a todo item
SimpleToDo master

Also, gave more control to editItemActivity to modify the item as needed.

+++ app/src/main/java/ravcode/com/mytodoapp/models/ToDoItem.java
@@ -42,7 +42,7 @@ public class ToDoItem extends Model {
         return item;
     }
 
-    public static ToDoItem getItemById(int id) {
+    public static ToDoItem getItemById(long id) {
         return new Select()
                 .from(ToDoItem.class)
                 .where("id = ?", id)
